float64 instead of int for length and time

+++ route.go
@@ -38,7 +38,7 @@ type RouteResponse struct {
 			Summary struct {
 				MaxLon float64 `json:"max_lon"`
 				MaxLat float64 `json:"max_lat"`
-				Time   int     `json:"time"`
+				Time   float64 `json:"time"`
 				Length float64 `json:"length"`
 				MinLat float64 `json:"min_lat"`
 				MinLon float64 `json:"min_lon"`
@@ -48,7 +48,7 @@ type RouteResponse struct {
 		Summary struct {
 			MaxLon float64 `json:"max_lon"`
 			MaxLat float64 `json:"max_lat"`
-			Time   int     `json:"time"`
+			Time   float64 `json:"time"`
 			Length float64 `json:"length"`
 			MinLat float64 `json:"min_lat"`
 			MinLon float64 `json:"min_lon"`
